move dedent to shared
