+++ e2e-pw/src/shared/python-runner/python-runner.ts
@@ -3,7 +3,7 @@ import dotenv from "dotenv";
 import fs from "fs";
 import os from "os";
 import path from "path";
-import { dedentPythonCode } from "src/oss/fixtures/dedent";
+import { dedentPythonCode } from "src/oss/utils/dedent";
 
 dotenv.config({ path: process.env.CI ? ".env.ci" : ".env.dev" });
